Add and implement reusable confirmation dialog hook

+++ src/components/data-table.tsx
@@ -13,6 +13,7 @@ import {
   TableHeader,
   TableRow,
 } from "@/components/ui/table";
+import { useConfirm } from "@/hooks/use-confirm";
 import {
   ColumnDef,
   ColumnFiltersState,
@@ -41,6 +42,11 @@ export function DataTable<TData, TValue>({
   onDelete,
   disabled,
 }: DataTableProps<TData, TValue>) {
+  const [ConfirmDialog, confirm] = useConfirm(
+    "Are you sure you?",
+    "Deleting accounts cannot be undone."
+  );
+
   const [sorting, setSorting] = React.useState<SortingState>([]);
   const [columnFilters, setColumnFilters] = React.useState<ColumnFiltersState>([]);
   const [rowSelection, setRowSelection] = React.useState({});
@@ -64,6 +70,7 @@ export function DataTable<TData, TValue>({
 
   return (
     <div>
+      <ConfirmDialog />
       <div className="flex flex-col gap-y-2 py-4 sm:flex-row sm:items-center">
         <Input
           placeholder={`Filter ${filterKey}...`}
@@ -75,9 +82,13 @@ export function DataTable<TData, TValue>({
           <Button
             disabled={disabled}
             className="w-full cursor-pointer border-none bg-rose-500 text-slate-100 shadow-none transition duration-300 ease-in-out outline-none hover:bg-rose-400 hover:text-slate-100 hover:shadow-md hover:shadow-rose-300/50 sm:ml-auto sm:w-auto"
-            onClick={() => {
-              onDelete(table.getFilteredSelectedRowModel().rows);
-              table.resetRowSelection();
+            onClick={async () => {
+              const ok = await confirm();
+
+              if (ok) {
+                onDelete(table.getFilteredSelectedRowModel().rows);
+                table.resetRowSelection();
+              }
             }}
           >
             <BsTrash className="size-4" />
